pyproject.toml
[project]
name = "dfm"
authors = [
    {name = "Josh Langsfeld", email = "josh.langsfeld@gmail.com"}
]
dependencies = [
    "pyyaml",
    "termcolor",
]
version = "0.1.0"

[project.scripts]
dfm = "dfm:main"
